fix: move key leakage unit test to CI
