+++ tests/litellm_utils_tests/test_health_check.py
@@ -637,58 +637,3 @@ async def test_image_generation_health_check_prompt(monkeypatch):
 
     assert len(health_check_calls) == 1
     assert health_check_calls[0]["prompt"] == override_prompt
-
-
-@pytest.mark.asyncio
-async def test_ahealth_check_failure_masks_raw_request_headers():
-    """
-    Security test: Verify that when ahealth_check() fails, the raw_request_headers
-    in raw_request_typed_dict are properly masked to prevent API key leaks.
-    
-    This tests the fix for the security vulnerability where Authorization headers
-    were being exposed in health check error responses.
-    """
-    # Use a model configuration that will fail (invalid endpoint)
-    test_api_key = "dapi-test-key-1234567890abcdef"
-    test_headers = {
-        "Authorization": f"Bearer {test_api_key}",
-        "Content-Type": "application/json",
-    }
-    
-    response = await litellm.ahealth_check(
-        model_params={
-            "model": "databricks/dbrx-instruct",
-            "api_base": "https://invalid-endpoint-that-will-fail.com/",
-            "api_key": test_api_key,
-            "headers": test_headers,
-        },
-        mode="chat",
-    )
-    
-    # Should have error and raw_request_typed_dict
-    assert "error" in response
-    assert "raw_request_typed_dict" in response
-    
-    raw_request_dict = response["raw_request_typed_dict"]
-    assert raw_request_dict is not None
-    assert isinstance(raw_request_dict, dict)
-    assert "raw_request_headers" in raw_request_dict
-    
-    headers = raw_request_dict["raw_request_headers"]
-    assert headers is not None
-    
-    # Security check: Authorization header should be masked, not show full key
-    if "Authorization" in headers:
-        auth_header = headers["Authorization"]
-        # Should be masked (e.g., "Be****90" or similar)
-        assert auth_header != f"Bearer {test_api_key}", "Authorization header must be masked"
-        assert auth_header != test_api_key, "API key must not appear in Authorization header"
-        # Masked headers typically have asterisks or are truncated
-        assert "*" in auth_header or len(auth_header) < len(f"Bearer {test_api_key}"), \
-            f"Authorization header should be masked but got: {auth_header}"
-    
-    # Content-Type should remain unmasked (not sensitive)
-    if "Content-Type" in headers:
-        assert headers["Content-Type"] == "application/json"
-    
-    print(f"Masked Authorization header: {headers.get('Authorization', 'NOT FOUND')}")
